# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Mini Project Management System API
  version: "1.0"
  description: |
    REST API cho hệ thống Mini Project Management System (Jira-like).
    Biên soạn dựa trên SRS-Mini-Project-Management-System.md và
    System-Architecture-Mini-Project-Management-System.md.

    ## Quy ước chung
    - Base path: `/api`
    - Mọi endpoint (trừ `/auth/login`, `/auth/refresh`) yêu cầu header
      `Authorization: Bearer <accessToken>`.
    - Response thành công luôn bọc trong `ApiResponseBase`:
      `{ "success": true, "message": null, "data": <payload> }`.
    - Response lỗi luôn theo `ErrorResponse`:
      `{ "success": false, "errorCode": "...", "message": "...", "details": [] }`.
    - Danh sách luôn hỗ trợ phân trang qua query `page` (mặc định 0) và
      `size` (mặc định 20), bọc trong `PageResponseBase` (NFR-04).
    - Định dạng ngày: `yyyy-MM-dd`. Định dạng thời gian: ISO-8601.
    - Phân quyền hai lớp theo System-Architecture mục 5.2: System Role
      (ADMIN/PM/DEV) kiểm tra ở Controller; Project Role (PM/DEV/TESTER,
      theo từng Project) kiểm tra ở Service — được ghi chú trong mô tả
      từng endpoint bên dưới.
servers:
  - url: /api
security:
  - bearerAuth: []
...
paths:
...
  /users:
    get:
      tags: [User]
      summary: Xem danh sách / tìm kiếm User (FR-USER-04, FR-USER-05)
      description: "Yêu cầu System Role: ADMIN"
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Size' }
        - name: keyword
          in: query
          schema: { type: string }
          description: Tìm theo Username/Full Name/Email
        - name: role
          in: query
          schema: { $ref: '#/components/schemas/SystemRole' }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/UserStatus' }
      responses:
        '200': { $ref: '#/components/responses/UserPageResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [User]
      summary: Tạo User (FR-USER-01)
      description: "Yêu cầu System Role: ADMIN. Password được băm BCrypt trước khi lưu."
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateUserRequest' }
      responses:
        '201': { $ref: '#/components/responses/UserResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /users/me:
    get:
      tags: [User]
      summary: Xem thông tin cá nhân (current user)
      description: "Yêu cầu đã đăng nhập (bất kỳ System Role nào). Trả về thông tin user của token hiện tại."
      responses:
        '200': { $ref: '#/components/responses/UserResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    put:
      tags: [User]
      summary: Cập nhật thông tin cá nhân (current user)
      description: "Yêu cầu đã đăng nhập (bất kỳ System Role nào). Cho phép cập nhật fullName, email, password. Không cho phép thay đổi role."
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateCurrentUserRequest' }
      responses:
        '200': { $ref: '#/components/responses/UserResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /users/{id}:
    get:
      tags: [User]
      summary: Xem chi tiết User
      parameters: [ { $ref: '#/components/parameters/IdPath' } ]
      responses:
        '200': { $ref: '#/components/responses/UserResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [User]
      summary: Cập nhật User (FR-USER-02)
      description: "Yêu cầu System Role: ADMIN"
      parameters: [ { $ref: '#/components/parameters/IdPath' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateUserRequest' }
      responses:
        '200': { $ref: '#/components/responses/UserResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /users/{id}/lock:
    patch:
      tags: [User]
      summary: Khoá / Mở khoá User (FR-USER-03)
      description: "Yêu cầu System Role: ADMIN"
      parameters: [ { $ref: '#/components/parameters/IdPath' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateUserStatusRequest' }
      responses:
        '200': { $ref: '#/components/responses/UserResponse' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ───────────────────────── PROJECT (FR-PROJ-01..03) ─────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, default: 0 }
    Size:
      name: size
      in: query
      schema: { type: integer, default: 20 }
    IdPath:
      name: id
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
  responses:
    BadRequest:
      description: Dữ liệu không hợp lệ / vi phạm ràng buộc nghiệp vụ
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Chưa đăng nhập hoặc token không hợp lệ/hết hạn
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Không đủ quyền (System Role hoặc Project Role)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Không tìm thấy tài nguyên
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    UserResponse:
      description: Thông tin User
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponseBase'
              - type: object
                properties:
                  data: { $ref: '#/components/schemas/UserDto' }
    UserPageResponse:
      description: Danh sách User có phân trang
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponseBase'
              - type: object
                properties:
                  data:
                    allOf:
                      - $ref: '#/components/schemas/PageResponseBase'
                      - type: object
                        properties:
                          content:
                            type: array
                            items: { $ref: '#/components/schemas/UserDto' }
...
  schemas:
...
    ApiResponseBase:
      type: object
      properties:
        success: { type: boolean, example: true }
        message: { type: string, nullable: true }
    PageResponseBase:
      type: object
      properties:
        totalElements: { type: integer, example: 128 }
        totalPages: { type: integer, example: 7 }
        pageNumber: { type: integer, example: 0 }
        pageSize: { type: integer, example: 20 }
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        errorCode: { type: string, example: VALIDATION_ERROR }
        message: { type: string, example: "Hour phải lớn hơn 0 và không vượt quá 24" }
        details:
          type: array
          items: { type: string }
          nullable: true

    # ---- Enum ----
    SystemRole: { type: string, enum: [ADMIN, USER] }
    UserStatus: { type: string, enum: [ACTIVE, LOCKED] }
...
    UserDto:
      type: object
      properties:
        id: { type: integer, format: int64 }
        employeeId: { type: string }
        username: { type: string }
        fullName: { type: string }
        email: { type: string, format: email }
        role: { $ref: '#/components/schemas/SystemRole' }
        status: { $ref: '#/components/schemas/UserStatus' }
    CreateUserRequest:
      type: object
      required: [employeeId, username, password, fullName, email, role]
      properties:
        employeeId: { type: string }
        username: { type: string }
        password: { type: string, format: password }
        fullName: { type: string }
        email: { type: string, format: email }
        role: { $ref: '#/components/schemas/SystemRole' }
    UpdateCurrentUserRequest:
      type: object
      properties:
        fullName: { type: string }
        email: { type: string, format: email }
        password: { type: string, format: password, minLength: 6 }
    UpdateUserRequest:
      type: object
      properties:
        fullName: { type: string }
        email: { type: string, format: email }
        role: { $ref: '#/components/schemas/SystemRole' }
    UpdateUserStatusRequest:
      type: object
      required: [status]
      properties:
        status: { $ref: '#/components/schemas/UserStatus' }

    # ---- Project ----
...